requirements.txt
fastapi[standard]==0.116.0
fastapi-crons==2.0.1
uvicorn==0.35.0
httpx==0.28.1
pydantic==2.11.7
Jinja2==3.1.6
slack-sdk==3.36.0
html-to-markdown==1.8.0
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
Markdown==3.8.2
google-cloud-logging==3.11.4